requirements.txt.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pymongo>=4.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0
mongomock>=4.0.0
